intl/l10n/test/dom/test_domloc_translateFragment.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test DOMLocalization.prototype.translateFragment</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript">
  "use strict";
  const { DOMLocalization } =
    ChromeUtils.import("resource://gre/modules/DOMLocalization.jsm");
  const { FluentBundle } =
    ChromeUtils.import("resource://gre/modules/Fluent.jsm");

  async function* mockGenerateMessages(locales, resourceIds) {
    const bundle = new FluentBundle(locales);
    bundle.addMessages("title = Hello World");
    bundle.addMessages("subtitle = Welcome to FluentBundle");
    yield bundle;
  }

  window.onload = async function() {
    SimpleTest.waitForExplicitFinish();

    const domLoc = new DOMLocalization(
      [],
      mockGenerateMessages
    );

    const frag = document.querySelectorAll("div")[0];
    const h1 = document.querySelectorAll("h1")[0];
    const p1 = document.querySelectorAll("p")[0];

    await domLoc.translateFragment(frag);
    is(h1.textContent, "Hello World");
    is(p1.textContent, "Welcome to FluentBundle");

    SimpleTest.finish();
  };
  </script>
</head>
<body>
  <div>
    <h1 data-l10n-id="title" />
    <p data-l10n-id="subtitle" />
  </div>
</body>
</html>
